backend/requirements.txt
Django==5.0.6
django-rest-framework==0.1.0
python-dotenv==1.0.1
requests==2.32.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pymemcache==4.0.0
mysqlclient==2.2.5
gunicorn==23.0.0
openai==1.52.2
drf-nested-routers==0.94.1